requirements.txt
Django==1.8.2
Markdown==2.6.2
django-filter==0.10.0
djangorestframework==3.1.3
requests==2.7.0